+++ apps/app/src/components/TemplateModal/TemplateModal.tsx
@@ -5,7 +5,7 @@ import React, {
 import assert from 'assert';
 
 import { Lang } from '@growi/core';
-import { useTemplateModal, type TemplateModalStatus } from '@growi/editor/src/stores/use-template';
+import { useTemplateModal, type TemplateModalStatus } from '@growi/editor/src/stores/use-template-modal';
 import {
   extractSupportedLocales, getLocalizedTemplate, type TemplateSummary,
 } from '@growi/pluginkit/dist/v4';
